Optimised existing WM detection and death message
The code to detect the existance of another window manager has been
optimised so that the program dies in the error handler if another
window manager is running rather than waiting to finish.

The death message and macro has been modified so that the file name and
line numbers are only included if the VERBOSE compiler directive is
defined.