First part of the settings reader
Added a function which reads through a configuration file to find lines
corresponding to settings for the program.
This function is called at the start of the program, before the bar
process is created and will currently cause the program to immediately
die.
Also modified the makefile to make use of the new default configuration
file in share/ when doing `make test`