Choose tags from bit in int
The bits inside an int now indicate which tag is selected, so there can
now be up to 32 tags on most systems. Bitwise and can be used to
determine is a window belongs to some tag, and bitwise or can be used to
set the tag of a window, etc.