Added ability to run commands in the same session
Commands specified in the configureation file can now be run in either
the same session as SXWM, or in a new session. The bar can be run in the
same session with 'run' so that when SXWM closes, so does the bar
process. But some daemon could be run with 'start' so that when SXWM
closes, the program keeps running in the background.