Improved die function
Replaced the 'die' macro with a function definition of the same now. The
function reads the basename of the program from the /proc filesystem, so
argv no longer needs to be preserved in a global variable by main. The
function also gives output in a single call to write(3) and does not
rely on libc file objects.